WT-2084 comments
