+++ packages/checkout/sdk/src/network/network.test.ts
@@ -6,7 +6,6 @@ import { Environment } from '@imtbl/config';
 import {
   getNetworkAllowList,
   getNetworkInfo,
-  isZkEvmChainId,
   switchWalletNetwork,
 } from './network';
 import { HttpClient } from '../api/http';
@@ -515,28 +514,6 @@ describe('network functions', () => {
           },
         ],
       });
-    });
-  });
-
-  describe('isZkEvmChainId', () => {
-    it('should return true if devnet zkEVM chain', () => {
-      const chainId = isZkEvmChainId(ChainId.IMTBL_ZKEVM_DEVNET);
-      expect(chainId).toBeTruthy();
-    });
-
-    it('should return true if testnet zkEVM chain', () => {
-      const chainId = isZkEvmChainId(ChainId.IMTBL_ZKEVM_TESTNET);
-      expect(chainId).toBeTruthy();
-    });
-
-    it('should return true if mainnet zkEVM chain', () => {
-      const chainId = isZkEvmChainId(ChainId.IMTBL_ZKEVM_MAINNET);
-      expect(chainId).toBeTruthy();
-    });
-
-    it('should return false if not zkEVM chain', () => {
-      const chainId = isZkEvmChainId(ChainId.SEPOLIA);
-      expect(chainId).toBeFalsy();
     });
   });
 });
