+++ packages/checkout/sdk/src/network/network.ts
@@ -209,7 +209,3 @@ export async function switchWalletNetwork(
     provider: newProvider,
   } as SwitchNetworkResult;
 }
-
-export const isZkEvmChainId = (chainId: ChainId) => chainId === ChainId.IMTBL_ZKEVM_DEVNET
-  || chainId === ChainId.IMTBL_ZKEVM_TESTNET
-  || chainId === ChainId.IMTBL_ZKEVM_MAINNET;
